+++ content/pages/client_page.yml
@@ -1,28 +1,1 @@
 clientsSectionTitle: Current and Past Clients
-testimonialsSectionTitle: null
-testimonials:
-  - quoteeName: Andrea Deveau
-    quote: Working with Alina has been an absolute pleasure and an incredibly
-      rewarding experience. Alina's creativity knows no bounds, and her
-      dedication to every project she takes on is truly inspiring. Look no
-      further if you're looking for someone who will elevate your efforts to new
-      heights. With Alina at the helm, success is not only possible but
-      inevitable.
-    organization: Deveau Burr Group
-  - quoteeName: Chris Bollinger
-    quote: Givher LLC is an exceptional organization. Alina treats every donor like
-      a VIP, ensuring we're informed, engaged, and genuinely appreciated. I feel
-      my donations are part of a larger movement towards positive change.
-    organization: Bollinger Government Affairs
-  - quoteeName: Assemblymember Evan Low
-    quote: Givher transformed our fundraising and event management! Their tailored
-      approach and top-notch support made a real difference for us. We're beyond
-      grateful! 🌟
-    organization: API Legislative Caucus Foundation
-  - quoteeName: John Robert Downs, Executive Director
-    quote: "\"Alina brought both expertise and warmth to our work with Givher. Her
-      understanding of California's legislative process was impressive, and
-      Givher’s strategic approach to logistics ensured that every detail of our
-      lobby day flowed seamlessly. Alina’s genuine interest in our mission made
-      us feel seen and fully supported at every step.\""
-    organization: Microdosing Collective
